Refactored Evaluation mechanism and started creating mechanism for adding new disease

+++ backend/PlantExpertSystem/src/main/java/pl/mleczko/PlantExpertSystem/Entity/Disease.java
@@ -20,7 +20,7 @@ public class Disease {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "disease_id")
-    private Integer diseaseId;
+    private Long diseaseId;
 
     @NonNull
     @Column(name = "name")
